examples/read_pointcloud/requirements.txt
-f ../../../prebuilt

bosdyn-client >= 3.1
numpy >= 1.16.4
open3d